docs: support Sphinx admonitions and refactor VitePress config
