+++ docs/Contributing.md
@@ -6,13 +6,13 @@ Engine modding is a complicated process which is pretty hard to pull off, but th
 
 You can observe how the stuff works by using the engine and note which other stuff influences the behavior, but sooner or later you would want to see the innards of that. This is usually done using such tools as disassemblers/decompilers ([IDA](https://www.hex-rays.com/products/ida/), [Ghidra](https://ghidra-sre.org/)) to decipher what is written in the binary (`gamemd.exe` in case of the binary) and debuggers ([Cheat Engine](https://www.cheatengine.org)'s debugger is pretty good for that) to trace how the binary works.
 
-::: tip Hint
+```{hint}
 Reverse-engineering is a complex task, but don't be discouraged, if you want to try your hands at it ask us in the Discord channel, we will gladly help 😄
-:::
+```
 
-::: info Note
+```{note}
 [Assembly language](https://www.cs.virginia.edu/~evans/cs216/guides/x86.html) and C++ knowledge, understanding of computer architecture, memory structure, OOP and compiler theory would certainly help.
-:::
+```
 
 ### Development
 
@@ -24,9 +24,9 @@ The repository includes a [Copilot instructions file](https://github.com/Phobos-
 
 We encourage contributors to try AI coding agents (such as GitHub Copilot in agent mode, Cursor, or similar tools) to assist with development tasks - writing hooks, reviewing code, generating documentation, and exploring the disassembly via [IDA Pro MCP](https://github.com/mrexodia/ida-pro-mcp). Agent skills (such as the [IDAPython skill](https://skills.sh/mrexodia/ida-pro-mcp/idapython)) can further extend agent capabilities with domain-specific knowledge. AI agents can significantly speed up routine work.
 
-::: info Note
+```{note}
 AI agents are a tool to assist development, but they are not perfect and can make mistakes. Always review and test any code generated or modified by an AI agent to ensure it meets the project's standards and works correctly. **You are responsible for the final code**, not the tool you use to write it.
-:::
+```
 
 ## Contributing changes to the project
 
@@ -44,21 +44,21 @@ If your change does not fit in standard criteria or too small that it doesn't ne
 
 If the pull request only changes documentation, start its title with `[Docs]` to skip the DLL build.
 
-::: tip Hint
+```{hint}
 Every pull request push trigger a nightly build for the latest pushed commit, so you can check the build status at the bottom of PR page, press `Show all checks`, go to details of a build run and get the zip containing built DLL and PDB (for your testers, f. ex.), or download a build from an automatically posted comment.
-:::
+```
 
-::: info Note
+```{note}
 You'd benefit from C++ experience, knowledge of programming patterns, common techniques etc. [Basic assembly knowledge](https://www.cs.virginia.edu/~evans/cs216/guides/x86.html) would help to correctly write the interaction with the memory where you hook at. Basic understanding of Git and GitHub is also needed.
-:::
+```
 
 ### Testing
 
 This is a job that any modder (and even sometimes player) can do. Look at a new feature or a change, try to think of all possible cases when it can work differently, try to think of any possible logic flaws, edge cases, unforeseen interactions or conditions etc., then test it according to your thoughts. Any bugs should be reported to issues section of this repo, if possible.
 
-::: warning Warning
+```{warning}
 **General stability** can only be achieved by extensive play-testing of new changes, both offline and online. Most modders have beta testing teams, so please, if you want the extension to be stable - contribute to that by having your testers play with the new features! Also the check-list below can help you identify issues quicker.
-:::
+```
 
 #### Testing check-list
 
@@ -69,9 +69,9 @@ This is a job that any modder (and even sometimes player) can do. Look at a new 
 - **Edge cases**. Those are the cases of some specific cases usually induced by some extreme parameter values (for example, vanilla game crashes on zero-size `PreviewPack` instead of not drawing it).
 - **Corner cases**. Those are similar to edge cases but are hard to reproduce and are usually induced by a combination of extreme parameter values.
 
-::: info Note
+```{note}
 Knowledge on how to mod YR and having an inquisitive mind, being attentive to details would help.
-:::
+```
 
 ### Writing docs
 
@@ -79,7 +79,7 @@ No explanation needed. If you fully understand how some stuff in Phobos works yo
 
 The docs are written in Markdown (which is dead simple, [learn MD in 60 seconds](https://commonmark.org/help/)). We use [VitePress](https://vitepress.dev/) to build the documentation site and [Read the Docs](https://readthedocs.org/) to host it.
 
-::: tip Hint
+```{hint}
 You can preview docs locally:
 
 1. `cd docs`
@@ -89,32 +89,32 @@ You can preview docs locally:
 For a production build, run `npm run build`.
 
 Chinese docs are built from `docs/locale/zh_CN/LC_MESSAGES/*.po` automatically by the VitePress PO locale plugin during `dev` and `build`.
-:::
+```
 
-::: tip Hint
+```{hint}
 You don't need to install Node.js, VitePress and npm packages locally just to see documentation changes - every pull request you make is built and served by Read the Docs automatically. Just like the nightly builds, scroll to the bottom of the PR page, press `Show all checks` and open the details of a Read the Docs build run to see the built documentation.
-:::
+```
 
 There are two ways to edit the docs.
 
 - **Edit from your PC**. Pretty much the same like what's described in [contributing changes section](#contributing-changes-to-the-project); the docs are located in the `docs` folder.
 - **Edit via online editor**. Navigate to the doc piece that you want to edit, press the button on the top right - and it will take you to the file at GitHub which you would need to edit (look for the pencil icon to the top right). Press it - the fork will be created and you'll edit the docs in your version of the repo (fork). You can commit those changes (preferably to a new branch) and make them into a pull request to main repo.
 
-::: info Note
+```{note}
 `README.md`, `CREDITS.md` and `LICENSE.md` are sourced from the repository root by the VitePress root pages plugin. Edit the files in the repository root instead.
-:::
+```
 
-::: info Note
+```{note}
 OK English grammar and understanding of docs structure would be enough. You would also need a GitHub account.
-:::
+```
 
 ### Providing media to showcase features
 
 Those would be used in docs and with a link to the respective mod as a bonus for the mod author. Please provide silent demonstration videos as WebM files.
 
-::: info Note
+```{note}
 Please, provide screenshots and WebM videos in their natural size and without excess stuff or length.
-:::
+```
 
 ### Promoting the work
