+++ docs/General-Info.md
@@ -10,9 +10,9 @@ There are three main types of Phobos builds:
 - _development builds_ - those are the builds which contain functionality that needs to be tested. They are numbered plainly starting from 0 and incrementing the number on each release. Mod authors still can include those versions with their mods if they want latest features, though we can't guarantee lack of bugs;
 - _nightly builds_ - bleeding edge versions which can include prototypes, proofs of concepts, scrapped features etc., in other words - we can't guarantee anything in those builds and they absolutely should NOT be used in mod releases and should only be used to help with development and testing.
 
-::: tip Hint
+```{hint}
 You can find the downloads for these versions on the document's [main page](/#downloads).
-:::
+```
 
 ### Disabling development build warning
